requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
chardet==5.2.0
    # via reportlab
cssselect2==0.8.0
    # via svglib
dj-database-url==3.0.0
    # via -r requirements.in
django==5.2.2
    # via
    #   dj-database-url
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-alias
    #   djangocms-versioning
    #   easy-thumbnails
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==5.0.1
    # via
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-frontend
    #   djangocms-link
    #   djangocms-versioning
django-entangled==0.6.2
    # via djangocms-frontend
django-filer==3.3.1
    # via
    #   -r requirements.in
    #   djangocms-frontend
django-formtools==2.5.1
    # via django-cms
django-fsm==2.8.2
    # via
    #   -r requirements.in
    #   djangocms-versioning
django-parler==2.3
    # via djangocms-alias
django-polymorphic==4.1.0
    # via django-filer
django-sekizai==4.1.0
    # via django-cms
django-treebeard==4.7.1
    # via django-cms
djangocms-admin-style==3.3.1
    # via django-cms
djangocms-alias==2.0.3
    # via -r requirements.in
djangocms-attributes-field==4.0.0
    # via
    #   djangocms-frontend
    #   djangocms-link
djangocms-frontend==2.1.2
    # via -r requirements.in
djangocms-link==5.0.1
    # via djangocms-frontend
djangocms-text==0.8.8
    # via -r requirements.in
djangocms-versioning==2.3.1
    # via -r requirements.in
easy-thumbnails[svg]==2.10
    # via
    #   django-filer
    #   djangocms-frontend
gunicorn==23.0.0
    # via -r requirements.in
lxml==5.4.0
    # via
    #   djangocms-text
    #   svglib
nh3==0.2.21
    # via djangocms-text
packaging==25.0
    # via
    #   django-cms
    #   djangocms-text
    #   djangocms-versioning
    #   gunicorn
pillow==11.2.1
    # via
    #   djangocms-text
    #   easy-thumbnails
    #   reportlab
psycopg2-binary==2.9.10
    # via -r requirements.in
reportlab==4.4.1
    # via
    #   easy-thumbnails
    #   svglib
sqlparse==0.5.3
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.4.0
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.14.0
    # via dj-database-url
webencodings==0.5.1
    # via
    #   cssselect2
    #   tinycss2

# The following packages are considered to be unsafe in a requirements file:
# setuptools
